recychbs-backend-main/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.10
botocore==1.35.10
certifi==2024.6.2
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.4
contourpy==1.1.1
cycler==0.12.1
DateTime==5.5
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-db-connection-pool==1.2.5
django-geoposition==0.3.0
django-heroku==0.3.1
django-jazzmin==3.0.0
django-storages==1.14.4
djangorestframework==3.15.2
filelock==3.14.0
Flask==3.0.3
fonttools==4.43.1
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.45.0
greenlet==3.1.1
gunicorn==23.0.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.0
MouseInfo==0.1.3
multidict==6.1.0
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.1
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
pipenv==2024.0.1
platformdirs==4.2.2
plyer==2.1.0
psycopg2==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyjokes==0.6.0
PyJWT==2.9.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
s3transfer==0.10.2
schedule==1.2.2
seaborn==0.13.0
setuptools==75.0.0
six==1.16.0
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.35
sqlparams==6.1.0
sqlparse==0.5.0
tabulate==0.9.0
twilio==9.3.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.0
virtualenv==20.26.2
Werkzeug==3.1.0
whitenoise==6.7.0
wikipedia==1.4.0
yarl==1.11.1
zope.interface==6.4.post2
